% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{adapt_timestep}
\alias{adapt_timestep}
\title{\enumerate{
\item Change of Timesteps
}}
\usage{
adapt_timestep(
  data_in,
  time_format = "\%d/\%m/\%Y \%H:\%M:\%S",
  time_column = 1,
  skip_hours = 24,
  data_columns = 3:ncol(data_in),
  timestep_out = 15
)
}
\arguments{
\item{data_in}{Input}

\item{time_format}{Time format according to strptime}

\item{time_column}{Number of the time column}

\item{skip_hours}{Hours to skip at the beginning}

\item{data_columns}{Columns with data (no time etc.)}

\item{timestep_out}{Timestep in minutes}
}
\description{
\enumerate{
\item Change of Timesteps
}
}
