% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{add_catchments}
\alias{add_catchments}
\title{Add Berlin Catchments to a map of Berlin}
\usage{
add_catchments(highlight_catchments = "", highlight_style = "shaded")
}
\arguments{
\item{highlight_catchments}{Character vector containing the catchments names
of catchments to be highlighted}

\item{highlight_style}{Either "shaded" (default) or a specific color, which
cab also be defined by \code{\link[=rgb]{rgb()}}}
}
\description{
Polygons are drawn in different types of gray and names are added, without
overlapping the catchment boundaries or any river
}
