% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{add_coloredRivers}
\alias{add_coloredRivers}
\title{Adds the colored rivers to the Berlin Map}
\usage{
add_coloredRivers(
  ext_riversList,
  sixBreaks,
  dataType = "time",
  LegendTitle,
  LegendLocation
)
}
\arguments{
\item{ext_riversList}{The list of Berlin rivers loaded with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}, where every table is extended with
\code{\link[=extend_riverTable]{extend_riverTable()}}}

\item{sixBreaks}{Breaks defining the lower limits of the categories. Must be
the same that were used in \code{\link[=extend_riverTable]{extend_riverTable()}}}

\item{dataType}{Is needed for the Legend. If "time" is used, it is assumed
that the first water quality category is between 0 and a value above 0, while
for the counting of events the first category is 0 only. If that is wanted
anything but "time" can be assigned.}

\item{LegendTitle}{String with title of the legend}

\item{LegendLocation}{Either "top" or "right" (outside the plot margin)}
}
\description{
Adds the colored rivers to the Berlin Map
}
