% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{add_constant_values}
\alias{add_constant_values}
\title{\enumerate{
\item Add constant values
}}
\usage{
add_constant_values(input_data, parameter_conversion)
}
\arguments{
\item{input_data}{Input}

\item{parameter_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=paraIDs]{paraIDs()}}}
}
\description{
\enumerate{
\item Add constant values
}
}
