% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{add_csoVol}
\alias{add_csoVol}
\title{Add Combined sewer overflow volume to Berlin map}
\usage{
add_csoVol(eventStats, sizeMax = 1e+05)
}
\arguments{
\item{eventStats}{Statistics data frame of one event, as saved by the
interface \code{\link[=iw_gerris_interface]{iw_gerris_interface()}}}

\item{sizeMax}{The maximum Volume in m³ to scale the point size. 100 000 is
a good compromise between small and large events. But if a single event
is described in more detail, this value can be adapted.}
}
\description{
Add Combined sewer overflow volume to Berlin map
}
