% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_continuousTimeIntervals.R
\name{adjust_time}
\alias{adjust_time}
\title{Equal time intervals}
\usage{
adjust_time(time_vector, time_interval)
}
\arguments{
\item{time_vector}{A POSIXct vector}

\item{time_interval}{Temporal resolution in seconds}
}
\value{
A POSIX vector with equally distanced points of time. The new
timesteps start on the hour and
}
\description{
Transforms the time vector, so that the every timestep is shifted to equal
equally distanced points of time
}
