% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_continuousTimeIntervals.R
\name{aggregate_measurements}
\alias{aggregate_measurements}
\title{Average data per predefined timesteps}
\usage{
aggregate_measurements(time_vector, data_vector, time_interval = 60 * 15)
}
\arguments{
\item{time_vector}{A POSIXct vector}

\item{data_vector}{A numeric vector, with data corresponding to the time_vector}

\item{time_interval}{Temporal resolution in seconds}
}
\value{
Data frame with POSIX time column "t" and data column "d
}
\description{
Creates a data frame with fitted time column and corresponding average
data values.
}
