% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{averaging_values}
\alias{averaging_values}
\title{3a. Integrated function in Change of Timesteps}
\usage{
averaging_values(values, timestep_in, timestep_out)
}
\arguments{
\item{values}{Vector of timeseries data}

\item{timestep_in, timestep_out}{Timesteps in the same unit}
}
\description{
Using a rule of three, the data is first scaled to a timestep of 1 minute and
then averaged for the required interval
}
