% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compareSites.R
\name{barPlot_site}
\alias{barPlot_site}
\title{Barplot comparing different Scenarios at one Site}
\usage{
barPlot_site(df_plot, siteName, barType)
}
\arguments{
\item{df_plot}{Data frame perpared by function \code{\link[=prepareBarplot]{prepareBarplot()}}}

\item{siteName}{One of the site names from the "Name" column of df_plot}

\item{barType}{One of "crit_ox", "comf_ox", "crit_events" and "neg_dev"}
}
\description{
Barplot comparing different Scenarios at one Site
}
\details{
Bar type "crit_ox" plots bars showing the amount of time the oxygen content
is below 0.5, 1, 1.5 and 2  mg/L.
Bar type "comf_ox" plots bars showing the amount of time the oxygen content
is below 3  mg/L (this was originally 5 mg/L in MiSa 1)
Bar type "crit_events" plots bars representing the amount critical oxygen
events (O2 < 1.5 mg/L)
Bar type "neg_dev" plots bars representing the negative deviation of oxygen
concentration compared to a non polluted site (MUE was defined in MiSa 1)
}
