% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{build_gerris_table}
\alias{build_gerris_table}
\title{\enumerate{
\item Reshaping to Gerris format
}}
\usage{
build_gerris_table(data_input)
}
\arguments{
\item{data_input}{Input}
}
\description{
\enumerate{
\item Reshaping to Gerris format
}
}
