% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{change_time_format}
\alias{change_time_format}
\title{\enumerate{
\item Change time to gerris format
}}
\usage{
change_time_format(input_data, new_format = "\%d.\%m.\%Y \%H:\%M")
}
\arguments{
\item{input_data}{Input}

\item{new_format}{Time format in strptime format}
}
\description{
\enumerate{
\item Change time to gerris format
}
}
