% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_run.R
\name{check_high_concentration}
\alias{check_high_concentration}
\title{Detailed look on concentration over the Gerris limit
special functions for more information about concentrations above gerris
concentration limit --> see concentrations and the according flow
iw_gerris_interace needs to be stopped at "5" to obtain the results table}
\usage{
check_high_concentration(result, parameterID_infoworks)
}
\arguments{
\item{result}{Table created with \code{\link[=iw_gerris_interface]{iw_gerris_interface()}}, stopped at step 5.}

\item{parameterID_infoworks}{The parameter to be checked. ID according to
infoworks}
}
\description{
Detailed look on concentration over the Gerris limit
special functions for more information about concentrations above gerris
concentration limit --> see concentrations and the according flow
iw_gerris_interace needs to be stopped at "5" to obtain the results table
}
\details{
special functions for more information about concentrations above gerris
concentration limit --> see concentrations and the according flow
iw_gerris_interace needs to be stopped at "5" to obtain the results table
}
