% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{check_parameter_range}
\alias{check_parameter_range}
\title{\enumerate{
\item Checking if the concentration lies within the gerris range
}}
\usage{
check_parameter_range(input_data, parameter_conversion)
}
\arguments{
\item{input_data}{Input}

\item{parameter_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=paraIDs]{paraIDs()}}}
}
\description{
\enumerate{
\item Checking if the concentration lies within the gerris range
}
}
