% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{closest_available_outlet}
\alias{closest_available_outlet}
\title{4a. Find closest outlet to the missing one}
\usage{
closest_available_outlet(missing_outlet, outlet_conversion)
}
\arguments{
\item{missing_outlet}{Infoworks upstream link ID of the Missing outlet}

\item{outlet_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=outletIDs]{outletIDs()}}}
}
\description{
4a. Find closest outlet to the missing one
}
