% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{correct_flow}
\alias{correct_flow}
\title{\enumerate{
\item Flow correction
}}
\usage{
correct_flow(
  input_data,
  data_columns = 3:ncol(input_data),
  flow_threshold = 0.003
)
}
\arguments{
\item{input_data}{input}

\item{data_columns}{columns data (no time etc.)}

\item{flow_threshold}{minimum flow in m3/s}
}
\description{
\enumerate{
\item Flow correction
}
}
