% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_deficit_assessment.R
\name{count_def_hours}
\alias{count_def_hours}
\title{Count hours of deficits}
\usage{
count_def_hours(data_vector, threshold, res)
}
\arguments{
\item{data_vector}{Numeric vector (with data in the same unit as the threshold)}

\item{threshold}{Numeric in the same unit as the data vector}

\item{res}{Temporal resolution of data in minutes}
}
\value{
A single Value (hours of deficits)
}
\description{
Count hours of deficits
}
