% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{data_per_outlet}
\alias{data_per_outlet}
\title{\enumerate{
\item Reshaping from parameter-tables to outlet-tables
}}
\usage{
data_per_outlet(input_data)
}
\arguments{
\item{input_data}{Input}
}
\description{
\enumerate{
\item Reshaping from parameter-tables to outlet-tables
}
}
