% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_misa_data.R
\name{decoupledCatchments}
\alias{decoupledCatchments}
\title{Dateiled information about catchments within one decoupling scenario}
\usage{
decoupledCatchments(decouplingScenario)
}
\arguments{
\item{decouplingScenario}{String as in colnames of the "outlet_decoupling.csv"
table, that can be loaded with \code{\link[=loadMisa_decouplingInfo]{loadMisa_decouplingInfo()}}}
}
\value{
A list of 2: 1) Character vector of all considered outlets of the scenario
(outled ID as used in Qsim)
2)Data frane with 5 columns: i) Catchment Names as defined in infoworks,
ii) Number of outlets per Catchment of all outlets, iii) number of outlets per
catchment of the decoupling scenario, iv) absolute decrease of outlets per
catchment compared to no decoupling, v) relative decrease of outlets
per catchment compared to no decoupling
}
\description{
Dateiled information about catchments within one decoupling scenario
}
