% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{extend_riverTable}
\alias{extend_riverTable}
\title{Adds Qualty Categories to the river table}
\usage{
extend_riverTable(river_table, qsim_misa_table, varName, sixBreaks)
}
\arguments{
\item{river_table}{Table of one of the rivers loaded with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}}

\item{qsim_misa_table}{Either the "hours", "events" or "neg_dev" data frame
of the mise assessment}

\item{varName}{One of the columns of the qsim_misa_table containing water
quality numeric information that should be plotted}

\item{sixBreaks}{Breaks defining the lower limits of the categories. The
values are included at the lower limit.}
}
\value{
The input river_table extended by three columns: 1) "value" containing the
exact value, 2) "quality" containing 6 different quality categories based on
the defined breaks, 3) "color" assigning th color for plotting
}
\description{
As a result, each location is linked to a water quality value.
}
\details{
The qsim_misa_table does not provide information for every location within
the river, since the local resolution of the qsim output differs from the
river course data. Values between two known river sites are interpolated.
The color scale cannot be changed. Since in all misa assessment parameters
low values refer to a good water quality, the colors range from green for low
values to red for high values.
}
