% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{find_positions}
\alias{find_positions}
\title{\enumerate{
\item small function to help creating a column with gerris outlet ID
}}
\usage{
find_positions(v_original, v_new)
}
\arguments{
\item{v_original}{Vector in old order}

\item{v_new}{Vector in new order}
}
\description{
\enumerate{
\item small function to help creating a column with gerris outlet ID
}
}
