% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_read_files.R
\name{finding_o2Column}
\alias{finding_o2Column}
\title{finding_o2Column}
\usage{
finding_o2Column(dataFrame, tryO2 = c("o2", "oxygen", "ox", "sauerstoff"))
}
\arguments{
\item{dataFrame}{The data frame where the column is searched}

\item{tryO2}{A vector with patterns possible patterns of columnnames with
oxygen data (not case sensitive)}
}
\value{
A vector with column numbers of oxygen columns
}
\description{
This function looks for the oxygen data column in a data frame by column
name
}
