% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_read_files.R
\name{finding_timestampColumns}
\alias{finding_timestampColumns}
\title{finding_timestampColumns}
\usage{
finding_timestampColumns(dataFrame)
}
\arguments{
\item{dataFrame}{The data frame where the column is searched}
}
\value{
A vector with column numbers of timestamp columns
}
\description{
This function looks for the timestamp column in a data frame by typical
timestamp symbols
}
