% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{get_concentrations}
\alias{get_concentrations}
\title{\enumerate{
\item calculating concentration from mass flows
}}
\usage{
get_concentrations(input_data, flowID_infoworks)
}
\arguments{
\item{input_data}{Input}

\item{flowID_infoworks}{Infoworks ID of flow}
}
\description{
\enumerate{
\item calculating concentration from mass flows
}
}
