% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{get_cso_stats}
\alias{get_cso_stats}
\title{Function to get statatistic per outlet}
\usage{
get_cso_stats(input_data, timestep_out, flow_only, outlet_conversion)
}
\arguments{
\item{input_data}{Input}

\item{timestep_out}{Timestep in strptime format}

\item{flow_only}{If TRUE BOB is set to 0}

\item{outlet_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=outletIDs]{outletIDs()}}}
}
\description{
Function to get statatistic per outlet
}
