% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{ids_from_colnames}
\alias{ids_from_colnames}
\title{ID from column name}
\usage{
ids_from_colnames(COLnames)
}
\arguments{
\item{COLnames}{Names of Columns (starting with "X" followed by the ID)}
}
\description{
ID from column name
}
