% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{integrate_missing_outlets}
\alias{integrate_missing_outlets}
\title{\enumerate{
\item Merge outlets that are not defined in Gerris with those that are
}}
\usage{
integrate_missing_outlets(data_input, outlet_conversion)
}
\arguments{
\item{data_input}{Input}

\item{outlet_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=outletIDs]{outletIDs()}}}
}
\description{
\enumerate{
\item Merge outlets that are not defined in Gerris with those that are
}
}
