% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_data.R
\name{interpolate_multipleNA}
\alias{interpolate_multipleNA}
\title{Linear interpolation for one or more missing values}
\usage{
interpolate_multipleNA(data_vector, max_na)
}
\arguments{
\item{data_vector}{Numeric vector of measurements (including NA values)}

\item{max_na}{the maximal number of NA values in a row to be interpolated}
}
\value{
A list containing the data vector with interpolated NA value as well as an
information about the amount of NA's interpolated
}
\description{
All sections of NA values that are smaller or equal as a defined maximal
number of NA's are interpolated
}
