% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{kn_correction}
\alias{kn_correction}
\title{\enumerate{
\item correction of Kjeldahl Nitrogen
}}
\usage{
kn_correction(input_data)
}
\arguments{
\item{input_data}{Input}
}
\description{
\enumerate{
\item correction of Kjeldahl Nitrogen
}
}
