% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misa_plot_functions.R
\name{load_berlin_rivers}
\alias{load_berlin_rivers}
\title{Reads package Data containing river courses of Berlin and assessed sites
by the Qsim model}
\usage{
load_berlin_rivers()
}
\value{
tables of all relevant rivers in misa combined in a list. Each table consists
of 4 columns: 1) longitued, 2) lattitude, 3) distance to fowollowing node in
km, 4) River km, 5) Corresponding Qsim output sitename, 6) Comment that can
be used for plots
}
\description{
Reads package Data containing river courses of Berlin and assessed sites
by the Qsim model
}
\details{
Geo-Data come from
Map data (c) OpenStreetMap contributors, https://www.openstreetmap.org
Extracts created by BBBike, https://extract.bbbike.org
osmium2shape by Geofabrik, https://geofabrik.de

Please read the OSM wiki how to use shape files.

https://wiki.openstreetmap.org/wiki/Shapefiles

shape file was created on: Fri 18 Jun 2021 03:58:25 PM UTC
GPS rectangle coordinates (lng,lat): 13.097,52.387 x 13.674,52.634
Script URL: https://extract.bbbike.org?sw_lng=13.097&sw_lat=52.387&ne_lng=13.674&ne_lat=52.634&format=shp.zip&city=Berlin&lang=de
Name of the area: Berlin
}
