% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loadBerlin.R
\name{load_focus_sites}
\alias{load_focus_sites}
\title{Loads river sites for comparing bar plots}
\usage{
load_focus_sites()
}
\value{
Table as defined in 'extdata/focus_sites' filtered for all rows
containing "y" in the "Auswahl"-column
}
\description{
Loads river sites for comparing bar plots
}
