% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{load_infoworks_output}
\alias{load_infoworks_output}
\title{\enumerate{
\item loading infoworks Output
}}
\usage{
load_infoworks_output(
  iw_output_folder,
  simulation_name,
  flow_only,
  parameter_conversion,
  outlet_conversion
)
}
\arguments{
\item{iw_output_folder}{Folder with infoworks output files}

\item{simulation_name}{Unique part of the simulation name}

\item{flow_only}{If TRUE only flow data frame is loaded}

\item{parameter_conversion}{The parameter conversion table from the package
can be loaded with \code{\link[=paraIDs]{paraIDs()}}}

\item{outlet_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=outletIDs]{outletIDs()}}}
}
\description{
\enumerate{
\item loading infoworks Output
}
}
