% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_misa_data.R
\name{loadMisa_decouplingInfo}
\alias{loadMisa_decouplingInfo}
\title{Loads table with decoupling information}
\usage{
loadMisa_decouplingInfo()
}
\value{
Table as defined in 'extdata/misa_data'
}
\description{
Loads table with decoupling information
}
