% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_misa_data.R
\name{loadMisa_events}
\alias{loadMisa_events}
\title{Loads critical rain events for scenario assessment}
\usage{
loadMisa_events()
}
\value{
Table as defined in 'extdata/misa_data' filtered
}
\description{
Loads critical rain events for scenario assessment
}
