% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_misa_data.R
\name{loadMisa_focus_sites}
\alias{loadMisa_focus_sites}
\title{Loads river sites for comparing bar plots}
\usage{
loadMisa_focus_sites()
}
\value{
Table as defined in 'extdata/misa_data' filtered for all rows
containing "y" in the "Auswahl"-column
}
\description{
Loads river sites for comparing bar plots
}
