% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{mapPlot_EventsNumber}
\alias{mapPlot_EventsNumber}
\title{Creates and stores a map Plot of all oxygen deficiencies Events}
\usage{
mapPlot_EventsNumber(
  BerlinRivers,
  df_aggr,
  savingPath,
  varName = "events",
  sixBreaks = c(-1, 0, 1, 3, 6, 10),
  dec = ","
)
}
\arguments{
\item{BerlinRivers}{Berlin rivers loaded from package with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}}

\item{df_aggr}{Aggregated misa assessment data frame}

\item{savingPath}{Directory where plot is going to be stored}

\item{varName}{Column name that is used for water quality categorisation}

\item{sixBreaks}{6 low limits of water quality categories}

\item{dec}{What should be used a decimal character? Default is "," since
the plots are in German language}
}
\description{
Creates and stores a map Plot of all oxygen deficiencies Events
}
