% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{mapPlot_EventsTime}
\alias{mapPlot_EventsTime}
\title{Creates and stores a map Plot of Oxygen deficiency time of all CSO Events}
\usage{
mapPlot_EventsTime(
  BerlinRivers,
  df_aggr,
  scenarioName = "",
  savingPath = "",
  varName = "hours.below_1.5",
  sixBreaks = c(0, 25, 50, 100, 200, 300),
  decoupling = "",
  dec = ",",
  highlight_catchments = NULL,
  highlight_style = "lightblue",
  english = FALSE,
  file_type = "png",
  size_scale = 1
)
}
\arguments{
\item{BerlinRivers}{Berlin rivers loaded from package with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}}

\item{df_aggr}{Aggregated misa assessment data frame}

\item{scenarioName}{A String, only used for the plot title and file name}

\item{savingPath}{Directory where plot is going to be stored. If left empty
(default) a new device is opened in RStudio for the plot.}

\item{varName}{Column name that is used for water quality categorisation}

\item{sixBreaks}{6 low limits of water quality categories}

\item{decoupling}{Character vector defining a decoupling scenario. Default
is "" for no scenario. Possible scenarios can be obtained by the column names}

\item{dec}{What should be used a decimal character? Default is "," since
the plots are in German language}

\item{highlight_catchments}{Character vector of catchments to highlight. If
NULL catchments will be highlighted based on decoupling scenario. Otherwise
this overwrites the decoupling information for catchments highlited.}

\item{highlight_style}{Either a color (name or \code{\link[=rgb]{rgb()}}) or "shaded" to use
the actual color as diagonal lines.}

\item{english}{If TRUE title of the legend is in english}

\item{file_type}{Either "png" or "wmf" to create a vector graphic (only
possible on windows computers)}

\item{size_scale}{Numeric value to increase (>1) or decrease (<1) the size
of the plot. Text size is not affected by th scale which is why small plots
have a relatively large text size.}
}
\description{
Creates and stores a map Plot of Oxygen deficiency time of all CSO Events
}
