% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{mapPlot_EventTime}
\alias{mapPlot_EventTime}
\title{Creates and stores a map Plot of Oxygen deficiency during CSO Event}
\usage{
mapPlot_EventTime(
  BerlinRivers,
  statFilesPath,
  dl_misa,
  event,
  scenarioName = "",
  savingPath = "",
  sizeMax = 1e+05,
  below = 1.5,
  sixBreaks = c(0, 0.5, 2, 4, 10, 20),
  decoupling = "",
  dec = ",",
  highlight_catchments = NULL,
  highlight_style = "lightblue"
)
}
\arguments{
\item{BerlinRivers}{Berlin rivers loaded from package with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}}

\item{statFilesPath}{Directory of all stat files}

\item{dl_misa}{Misa assessment list of all events}

\item{event}{Event number to be plotted.}

\item{scenarioName}{A String, only used for the plot title and file name}

\item{savingPath}{Directory where plot is going to be stored. If left empty
(default) a new device is opened in RStudio for the plot.}

\item{sizeMax}{Maximum size of CSO volumne. Default 100000 is a good
compromise between large and small CSO events}

\item{below}{Oxygen concentration limit used for the plot (mg/L)}

\item{sixBreaks}{6 low limits of water quality categories}

\item{decoupling}{Character vector defining a decoupling scenario. Default
is "" for no scenario. Possible scenarios can be obtained by the column names
of \code{\link[=loadMisa_decouplingInfo]{loadMisa_decouplingInfo()}}.}

\item{dec}{What should be used a decimal character? Default is "," since
the plots are in German language}

\item{highlight_catchments}{Character vector of catchments to highlight. If
NULL catchments will be highlighted based on decoupling scenario. Otherwise
this overwrites the decoupling information for catchments highlighted.}

\item{highlight_style}{Either a color (name or \code{\link[=rgb]{rgb()}}) or "shaded" to use
the actual color as diagonal lines.}
}
\description{
Creates and stores a map Plot of Oxygen deficiency during CSO Event
}
