% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_filter_data.R
\name{misa_filter_data}
\alias{misa_filter_data}
\title{Filter MiSa Dataframe}
\usage{
misa_filter_data(
  dataFrame,
  sites = "",
  tBeg = min(dataFrame$posixDateTime, na.rm = TRUE),
  tEnd = max(dataFrame$posixDateTime, na.rm = TRUE)
)
}
\arguments{
\item{dataFrame}{Data frame loaded by a MiSa function (see details)}

\item{tBeg}{POSIX-Value with a start time of the observeration interval}

\item{tEnd}{POSIX-Value with an end time of the observeration interval}

\item{site}{Names of considered sites, written in the site columne of a MiSa
Dataframe}
}
\value{
A filtered data frame with the same columns as the input data frame
}
\description{
Filters the loaded data frame by sites and time
}
\details{
The name of the site column must be "site", the name of the timestamp column
must be "posixDateTime".
The best way is to load the oxygen data  with one of the following
functions: \code{\link[=read_misa_oneSite]{read_misa_oneSite()}},\code{\link[=read_misa_multipleSites]{read_misa_multipleSites()}} or
\code{\link[=read_misa_files]{read_misa_files()}}.
}
