% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_prepare_data.R
\name{misa_prepare_data}
\alias{misa_prepare_data}
\title{Prepare MiSa Data for MiSa Assessment}
\usage{
misa_prepare_data(df_MiSa, res = 15, max_na_interpolation = 60/res)
}
\arguments{
\item{df_MiSa}{Data frame loaded with one of the MiSa Load functions}

\item{res}{Temporal resolution in minutes}

\item{max_na_interpolation}{Maximal numbers of NA values in a row to be
interpolated. The default is one hour without measurements. Number of NA
depneds on the temporal resolution (60 / res)}
}
\value{
List with data frames per site, that is ready for MiSa Assessmen. Additional
information is printed about the number of interpolated NA values. If there
are many NA values that are not interpolated it is probably due to the fact
of no measurements during winter.
}
\description{
Timestamps are adapted, oxygen data is interpolated, it is filtered for
summmer months
}
