% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compareSites.R
\name{misaBars}
\alias{misaBars}
\title{Draws all bars within the function \code{\link[=barPlot_site]{barPlot_site()}}}
\usage{
misaBars(df_site, y_var, Color)
}
\arguments{
\item{df_site}{data frame of all secnarios considered at one side}

\item{y_var}{Variable column name in df_site that is used for y-values}

\item{Color}{Color of bars}
}
\description{
Draws all bars within the function \code{\link[=barPlot_site]{barPlot_site()}}
}
