% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MisaColor}
\alias{MisaColor}
\title{Color palette for MiSa Plots}
\format{
Named vector of length 6
}
\source{
Own selection
}
\usage{
MisaColor
}
\description{
Contains 6 different colors representing a heat scale from dark turqoise
for "perfect" to dark pink for "very serious"
}
\keyword{datasets}
