% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{outletIDs}
\alias{outletIDs}
\title{Loads the package outlet ID table}
\usage{
outletIDs()
}
\description{
Loads the package outlet ID table
}
\details{
Columns:
outlet_id: Node ID of the outlet within the infoworks model
upstream_link_id: Infoworks ID of the link leading to outlet (this must be
usedfor mass flow information)
gerris_id: "Randbedingung" used in gerris (provided by Schumacher,
Ingenieurbüro für Wasser und Umwelt)
surface_water: name of the surface water the outlet discharges into
water_body_km: the water body km where the outlet is located
y_coordinate and x_coordanite: outlet location in soldner coordinates for
Berlin
all lat and lon columns: coordinates transformed from Soldner to UTM-Format
(ETRS89 DREF91)
comment: possibility to add comments (column not used by r-script)
}
