% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_misa_data.R
\name{paraIDs}
\alias{paraIDs}
\title{Loads the parameter ID table}
\usage{
paraIDs()
}
\description{
Loads the parameter ID table
}
\details{
Columns:
parameter: full German Parameter Name
id_infoworks: parameter ID in Infoworks (mf = mass flow)
unit_infoworks: unit of the infoworks output (mass flow)
constant_value: parameters that are not simulated by infoworks and are set
to a constant value
unit is the same as in gerris (see column "unit_gerris")
id_gerris: parameter ID in Gerris
unit_gerris: unit needed for Gerris input (concentration)
min_gerris and max_gerris: tolerated range of parameter in Gerris model
comment: possibility to add comments (column not used by r-script)
}
