% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{postText}
\alias{postText}
\title{Text written and status bar drawn after data process step}
\usage{
postText(x, step)
}
\arguments{
\item{x}{Character describing the process step}

\item{step}{Number of process step}
}
\description{
Text written and status bar drawn after data process step
}
