% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_misa_data.R
\name{prepare_misa_data}
\alias{prepare_misa_data}
\title{Prepare MiSa Data for MiSa Assessment}
\usage{
prepare_misa_data(df_MiSa, res = 15, max_na_interpolation = 4)
}
\arguments{
\item{df_MiSa}{Data frame loaded with one of the MiSa Load functions}

\item{res}{Temporal resolution in minutes}

\item{max_na_interpolation}{Maximal numbers of NA values in a row to be
interpolated}
}
\value{
List with data frames per site, that is ready for MiSa Assessment
}
\description{
Timestamps are adapted, oxygen data is interpolated, it is filtered for
summmer months
}
