% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compareSites.R
\name{prepareBarplot}
\alias{prepareBarplot}
\title{Loads different Misa Assessment output files and writes them into one table}
\usage{
prepareBarplot(scenarioNames, scenarioPath)
}
\arguments{
\item{scenarioNames}{Names of the misa output files (.Rdata-Files)}

\item{scenarioPath}{Path of all misa output files to be compared}
}
\value{
One table wit all misa output data at the focus sites
}
\description{
Comparison of the aggregated misa files at specified focus river sites
}
