% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{preText}
\alias{preText}
\title{Text written before data process step}
\usage{
preText(x)
}
\arguments{
\item{x}{Character describing the process step}
}
\description{
Text written before data process step
}
