% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_Qsim_output.R
\name{QSIM_get_flow}
\alias{QSIM_get_flow}
\title{Read Qsim output and saves the flow at designated river sites}
\usage{
QSIM_get_flow(qsim_output_path, qsim_fileName, save_path)
}
\arguments{
\item{qsim_output_path}{path where the output file from Gsim is stored}

\item{qsim_fileName}{Filename of the Qsim output file (including .csv)}

\item{save_path}{path where flow table is saved}
}
\description{
Read Qsim output and saves the flow at designated river sites
}
