% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Qsim_output.R
\name{QSIM_prepare_for_tool}
\alias{QSIM_prepare_for_tool}
\title{Read Qsim output and save in required format for misa assessment tool}
\usage{
QSIM_prepare_for_tool(
  qsim_output_path,
  qsim_fileName,
  misa_tool_input_path,
  output_fileName,
  return_table = FALSE
)
}
\arguments{
\item{qsim_output_path}{path where the output file from Gsim is stored}

\item{qsim_fileName}{Filename of the Qsim output file (inclusive .csv)}

\item{misa_tool_input_path}{input Folder for Misa Tool (see details)}

\item{output_fileName}{Filename of the created table (inclusive .csv)}

\item{return_table}{if TRUE, the saved table is returned}
}
\value{
The saved table contains a timestamp column in Central European Time and
several oxygen columns in mg/L
}
\description{
This function combines the Qsim site information of the first 3 columns
to unique sitenames and spreads the Qsim output table into a table with one
time column and several oxygen columns (one per site)
}
\details{
The Misa Tool input folder as defined in parameter 'misa_tool_input_path' must
contain the two folder "files_per_site" and "site_per_file". The output
table will be stored in the latter.
Saves table as CSV File ";"-separated if parameter return_table is set to TRUE
}
