% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_read_files.R
\name{read_misa_files}
\alias{read_misa_files}
\title{Read MiSa Files}
\usage{
read_misa_files(input_path)
}
\arguments{
\item{input_path}{This is the directory where the two folders "files_per_site"
and "sites_per_file" are located}
}
\value{
A Data frame with 3 columen: Timestamp, Oxygen data and Site name
}
\description{
This function combines the functions read_misa_oneSite and
read_misa_multipleSites and is strictly bound the the misa folder structure
}
\details{
All csv files from both folders will be loaded. They must contain a timestamp
column and an oxygen column. The timestamp column is identified
automatically, by looking for a column where the entries contain ":" and one
of the date separating symbols ".", "/" or "-".. The oxygen column is
found by its colname. It should conatin "O2", "o2", "Oxygen", "oxygen", "ox",
"Ox", "Sauerstoff" or "sauerstoff".

For "file_per_site" files: all letters in the filename before the first "_" are
used for the sitename.
For "sites_per_file" files: all column names will be used for site names.
Thus, all columns except the timestamp must be oxygan concentrations at
different sites
}
