% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misa_read_files.R
\name{read_misa_multipleSites}
\alias{read_misa_multipleSites}
\title{read_misa_multipleSites}
\usage{
read_misa_multipleSites(path, file)
}
\arguments{
\item{path}{The path to the file}

\item{file}{Filename (including ".csv" Ending)}
}
\value{
A Data frame with 3 columns: Timestamp, Oxygen data and Site name
}
\description{
This function reads csv tables with one timestamp column and several
oxygen data columns. Where the colnames refer to the sites of measurements
}
