% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_read_files.R
\name{read_misa_oneSite}
\alias{read_misa_oneSite}
\title{read_misa_oneSite}
\usage{
read_misa_oneSite(path, file, siteID)
}
\arguments{
\item{path}{The path to the file}

\item{file}{Filename (including ".csv" Ending)}

\item{siteID}{a character vector specifying the site name}
}
\value{
A Data frame with 3 columns: Timestamp, Oxygen data and Site name
}
\description{
This function reads csv tables with one timestamp column and one
oxygen data column
}
