% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{rename_IDs}
\alias{rename_IDs}
\title{\enumerate{
\item Renaming parameter IDs
}}
\usage{
rename_IDs(input_data, parameter_conversion)
}
\arguments{
\item{input_data}{Input}

\item{parameter_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=paraIDs]{paraIDs()}}}
}
\description{
\enumerate{
\item Renaming parameter IDs
}
}
