% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{same_inarow}
\alias{same_inarow}
\title{Repeating values in a row within a Vector}
\usage{
same_inarow(v, NA_treatment = "NA")
}
\arguments{
\item{v}{A character, factor or numeric vector}

\item{NA_treatmen}{How to treat NA values within the vector. Defaul is NA,
It can also be "0" (FALSE) or "1" (TRUE)}
}
\value{
A data frame with four columns: Value (-> listed value of the input vector),
Repeats (times it is repeated in a row), starts_at (start position),
ends_at (end position).
}
\description{
Describes the Values of a vector the times they are repeated and the
start and end position of those values
}
