% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{save_cso_stats}
\alias{save_cso_stats}
\title{16b. Save interface output (stats)}
\usage{
save_cso_stats(input_data, gerris_input_folder, output_filename)
}
\arguments{
\item{input_data}{Stats}

\item{gerris_input_folder}{Output folder where data is saved}

\item{output_filename}{Filename}
}
\description{
16b. Save interface output (stats)
}
