% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{save_gerris_input}
\alias{save_gerris_input}
\title{\enumerate{
\item Save interface output
}}
\usage{
save_gerris_input(input_data, gerris_input_folder, output_filename)
}
\arguments{
\item{input_data}{Input}

\item{gerris_input_folder}{Output folder where data is saved}

\item{output_filename}{Filename}
}
\description{
\enumerate{
\item Save interface output
}
}
