% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_Qsim_output.R
\name{siteInfo_from_QsimName}
\alias{siteInfo_from_QsimName}
\title{Get River, Section and location information from Qsim site names}
\usage{
siteInfo_from_QsimName(misa_assessment)
}
\arguments{
\item{misa_assessment}{The assessment of a Qsim simulation}
}
\description{
Get River, Section and location information from Qsim site names
}
