% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiSa_filter_data.R
\name{SummerMonths}
\alias{SummerMonths}
\title{Breaking down POSIX into months and year and filter by month}
\usage{
SummerMonths(df, time_column = "t", months = 5:9)
}
\arguments{
\item{df}{data frame with a POSIX time column}

\item{time_column}{Name or number of the time column}

\item{months}{the number of the months that should be kept. The default is
5:9 which are the important months for oxygen deficits in Berlin caused by
CSOs}
}
\value{
The filteterd input data frame with months and year column
}
\description{
Adds month and year column to data frame and filters by month. The addition
of the year column is important for the following MiSa Assessment
}
