% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_functions.R
\name{suppress_flow}
\alias{suppress_flow}
\title{2a. suppress outlet Overflow}
\usage{
suppress_flow(input_data, outlets_to_suppress, outlet_conversion)
}
\arguments{
\item{input_data}{input}

\item{outlets_to_suppress}{Character vector defining the Infoworks outlet
IDs of all outlets to be suppressed (->set to 0)}

\item{outlet_conversion}{The outlet conversion table from the package
can be loaded with \code{\link[=outletIDs]{outletIDs()}}}
}
\description{
2a. suppress outlet Overflow
}
