% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misa_assessment_functions.R
\name{yearly_crit_Events}
\alias{yearly_crit_Events}
\title{MiSa Assessment: Yearly Numbers of deficits}
\usage{
yearly_crit_Events(
  dataFrame,
  res = 15,
  seperating_hours = 5 * 24,
  deficiency_hours = 0.25,
  thresholds = 1.5,
  max_missing = 25,
  use_recovery_value = FALSE,
  recovery_value = NULL
)
}
\arguments{
\item{dataFrame}{MiSa Dataframe: with columne "d": oxygen data, "year": year}

\item{res}{Temporal resolution of oxygen data in minutes}

\item{seperating_hours}{TODO: describe (also: should be "separating_hours"
with "a", not "e")}

\item{deficiency_hours}{TODO: describe}

\item{thresholds}{Oxygen threshold values used for the assessment in mg/L}

\item{max_missing}{The maximal allowed percent of missing oxygen data. If NA
Values exceed this number, hours below thresholds are set to NA}

\item{use_recovery_value}{TODO: describe}

\item{recovery_value}{TODO: describe}
}
\value{
Data frane with rows per year and columns per threshold as well es
for missing data
}
\description{
Counting the events below threshold values on a yearly basis
}
