% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misa_assessment_functions.R
\name{yearly_deficiency_time}
\alias{yearly_deficiency_time}
\title{MiSa Assessment: Yearly hours of deficits}
\usage{
yearly_deficiency_time(
  dataFrame,
  res = 15,
  thresholds = c(0.5, 1, 1.5, 2, 5),
  max_missing = 25
)
}
\arguments{
\item{dataFrame}{MiSa Dataframe: with columne "d": data, "year": year}

\item{res}{Temporal resolution of oxygen data in minutes}

\item{thresholds}{Oxygen threshold values used for the assessment in mg/L}

\item{max_missing}{The maximal allowed percent of missing oxygen data. If NA
Values exceed this number, hours below thresholds are set to NA}
}
\value{
Data frane with rows per year and columns per threshold as well es
for missing data
}
\description{
Counting the hours on a yearly basis below threshold values
}
