% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_final_functions.R
\name{yearly_negative_deviation}
\alias{yearly_negative_deviation}
\title{Negative deviation from a reference site}
\usage{
yearly_negative_deviation(dataFrame, oxygen_ref, max_missing = 25)
}
\arguments{
\item{dataFrame}{MiSa Dataframe: with columne "d": oxygen data, "year": year}

\item{oxygen_ref}{The course of oxygen of the reference}

\item{max_missing}{The maximal allowed percent of missing oxygen data. If NA
Values exceed this number, hours below thresholds are set to NA}
}
\value{
Data frane with negative deviation per year
}
\description{
Functions cumulates the negative deviation (lower O2-Concentrations) compared
to a reference site without (significant) urban pollution
}
