% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{add_catchments}
\alias{add_catchments}
\title{Add Berlin Catchments to a map of Berlin}
\usage{
add_catchments()
}
\description{
Polygons are drawn in different types of gray and names are added, without
overlapping the catchment boundaries or any river
}
