% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_Qsim_output.R
\name{df_aggr}
\alias{df_aggr}
\title{Aggregates the MiSa assessment of complete event series}
\usage{
df_aggr(dl_misa)
}
\arguments{
\item{dl_misa}{List of events}
}
\description{
Aggregates the MiSa assessment of complete event series
}
