% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapBerlin.R
\name{mapPlot_EventTime}
\alias{mapPlot_EventTime}
\title{Creates and stores a map Plot of Oxygen deficiency during CSO Event}
\usage{
mapPlot_EventTime(
  BerlinRivers,
  statFilesPath,
  dl_misa,
  event,
  savingPath,
  sizeMax = 1e+05,
  varName = "below_1.5",
  sixBreaks = c(0, 0.5, 2, 4, 10, 20),
  dec = ","
)
}
\arguments{
\item{BerlinRivers}{Berlin rivers loaded from package with
\code{\link[=load_berlin_rivers]{load_berlin_rivers()}}}

\item{statFilesPath}{Directory of all stat files}

\item{dl_misa}{Misa assessment list of all events}

\item{event}{Event number to be plotted}

\item{savingPath}{Directory where plot is going to be stored}

\item{sizeMax}{Maximum size of CSO volumne. Default 100000 is a good
compromise between large and small CSO events}

\item{varName}{Column name that is used for water quality categorisation}

\item{sixBreaks}{6 low limits of water quality categories}

\item{dec}{What should be used a decimal character? Default is "," since
the plots are in German language}
}
\description{
Creates and stores a map Plot of Oxygen deficiency during CSO Event
}
