% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compareSites.R
\name{prepareBarplot}
\alias{prepareBarplot}
\title{Loads a set of Misa Assessment output files and writes them into one table}
\usage{
prepareBarplot(rdata_files, scenario_names, qsim_focus_sites)
}
\arguments{
\item{rdata_files}{A list of filnames (including filt paths) of the MiSa
assessment output files. (.RData-files containing "df_aggr" and "dl_misa")}

\item{scenario_names}{A character vector of scenario names correspronding to
the list of RData files.}

\item{qsim_focus_sites}{A vector of Qsim IDs of the sites to be compared}
}
\value{
One table wit all Misa output data at the focus sites
}
\description{
Comparison of the aggregated misa files at specified focus river sites
}
